+++ pkg/yatee/samples/compose.yatee
