+++ pkg/yatee/samples/settings.yaml
